requirements.txt
chainer<2.0.0
opencv-python
pillow
scikit-image
